fix: align live field indentation, add debounce for live updates, fix type mismatch
Fix inconsistent indentation of `live: false` across block definitions,
change audiogain amplification type from f32 to f64 for consistency with
PropertyValue::Float, add 80ms debounce on live property API calls to
prevent flooding during slider drags, and fix misplaced doc comments.

+++ backend/src/blocks/builtin/audiogain.rs
@@ -95,7 +95,7 @@ impl BlockBuilder for AudioGainBuilder {
                 _ => None,
             })
             .unwrap_or(false);
-        let amplification: f32 = if inverted { -1.0 } else { 1.0 };
+        let amplification: f64 = if inverted { -1.0 } else { 1.0 };
 
         info!(
             "AudioGain properties: gain_db={}, gain_linear={}, mute={}, inverted={}",
